requirements.txt
honcho
celery
redis
fastapi
python-socketio
uvicorn